test: Improve tests to better adhere to the pattern
