+++ tools/serverpod_cli/test/analytics/generate_tracker_test.dart
@@ -15,97 +15,97 @@ const _snapshot = ProtocolAnalyticsSnapshot(
 );
 
 void main() {
-  group('Given a GenerateTracker, ', () {
-    late GenerateTracker tracker;
-    late RecordingAnalytics recording;
-    late String serverDir;
+  late GenerateTracker tracker;
+  late RecordingAnalytics recording;
+  late String serverDir;
 
-    setUp(() async {
-      tracker = GenerateTracker(debounceDuration: Duration.zero);
-      recording = RecordingAnalytics();
-      initializeCliAnalytics(
-        CliAnalytics(analytics: recording)..enabled = true,
+  setUp(() async {
+    tracker = GenerateTracker(debounceDuration: Duration.zero);
+    recording = RecordingAnalytics();
+    initializeCliAnalytics(
+      CliAnalytics(analytics: recording)..enabled = true,
+    );
+
+    // Anchor git-dir resolution inside the sandbox so metadata never escapes
+    // to an ambient repo (e.g. a stray /tmp/.git).
+    await d.dir('.git', [d.file('config', '')]).create();
+    await d.dir('tracker_project', [
+      d.dir('myapp_server', [d.file('pubspec.yaml', 'name: myapp_server')]),
+    ]).create();
+
+    serverDir = p.join(d.sandbox, 'tracker_project', 'myapp_server');
+  });
+
+  tearDown(() => initializeCliAnalytics(CliAnalytics.disabled()));
+
+  test(
+    'Given two incremental generation runs in the same burst, '
+    'when the debounce period ends, '
+    'then one event reports their run count and mean duration.',
+    () async {
+      final config = buildAnalyticsTestConfig(serverDir);
+
+      tracker.recordIncrementalRun(
+        config: config,
+        success: true,
+        duration: const Duration(milliseconds: 100),
+        snapshot: _snapshot,
+      );
+      tracker.recordIncrementalRun(
+        config: config,
+        success: true,
+        duration: const Duration(milliseconds: 300),
+        snapshot: _snapshot,
       );
 
-      // Anchor git-dir resolution inside the sandbox so metadata never escapes
-      // to an ambient repo (e.g. a stray /tmp/.git).
-      await d.dir('.git', [d.file('config', '')]).create();
-      await d.dir('tracker_project', [
-        d.dir('myapp_server', [d.file('pubspec.yaml', 'name: myapp_server')]),
-      ]).create();
+      await Future<void>.delayed(const Duration(milliseconds: 50));
 
-      serverDir = p.join(d.sandbox, 'tracker_project', 'myapp_server');
-    });
+      expect(recording.events, ['cli.generate']);
 
-    tearDown(() => initializeCliAnalytics(CliAnalytics.disabled()));
+      final properties = recording.properties.single;
+      expect(properties['is_watch_mode'], isTrue);
+      expect(properties['incremental_run_count'], 2);
+      expect(properties['incremental_avg_duration_ms'], 200);
+      expect(
+        properties.containsKey('oneshot_duration_ms'),
+        isFalse,
+        reason: 'One-shot and watch timings must stay mutually exclusive.',
+      );
+    },
+  );
 
-    test(
-      'when two incremental runs are debounced, '
-      'then one cli.generate event carries the run count and mean duration.',
-      () async {
-        final config = buildAnalyticsTestConfig(serverDir);
+  test(
+    'Given a flushed generation burst followed by a new run, '
+    'when the second debounce period ends, '
+    'then the second event reports only the new run.',
+    () async {
+      final config = buildAnalyticsTestConfig(serverDir);
 
-        tracker.recordIncrementalRun(
-          config: config,
-          success: true,
-          duration: const Duration(milliseconds: 100),
-          snapshot: _snapshot,
-        );
-        tracker.recordIncrementalRun(
-          config: config,
-          success: true,
-          duration: const Duration(milliseconds: 300),
-          snapshot: _snapshot,
-        );
+      tracker.recordIncrementalRun(
+        config: config,
+        success: true,
+        duration: const Duration(milliseconds: 100),
+        snapshot: _snapshot,
+      );
+      await Future<void>.delayed(const Duration(milliseconds: 50));
 
-        await Future<void>.delayed(const Duration(milliseconds: 50));
+      tracker.recordIncrementalRun(
+        config: config,
+        success: false,
+        duration: const Duration(milliseconds: 400),
+        snapshot: _snapshot,
+      );
+      await Future<void>.delayed(const Duration(milliseconds: 50));
 
-        expect(recording.events, ['cli.generate']);
-
-        final properties = recording.properties.single;
-        expect(properties['is_watch_mode'], isTrue);
-        expect(properties['incremental_run_count'], 2);
-        expect(properties['incremental_avg_duration_ms'], 200);
-        expect(
-          properties.containsKey('oneshot_duration_ms'),
-          isFalse,
-          reason: 'One-shot and watch timings must stay mutually exclusive.',
-        );
-      },
-    );
-
-    test(
-      'when a new burst follows a flushed one, '
-      'then the second event counts only the new runs.',
-      () async {
-        final config = buildAnalyticsTestConfig(serverDir);
-
-        tracker.recordIncrementalRun(
-          config: config,
-          success: true,
-          duration: const Duration(milliseconds: 100),
-          snapshot: _snapshot,
-        );
-        await Future<void>.delayed(const Duration(milliseconds: 50));
-
-        tracker.recordIncrementalRun(
-          config: config,
-          success: false,
-          duration: const Duration(milliseconds: 400),
-          snapshot: _snapshot,
-        );
-        await Future<void>.delayed(const Duration(milliseconds: 50));
-
-        expect(recording.events, ['cli.generate', 'cli.generate']);
-        expect(recording.properties.last['incremental_run_count'], 1);
-        expect(recording.properties.last['incremental_avg_duration_ms'], 400);
-        expect(recording.properties.last['generation_succeeded'], isFalse);
-        expect(
-          recording.properties.last['num_generate_calls'],
-          2,
-          reason: 'The metadata counter increments once per emitted event.',
-        );
-      },
-    );
-  });
+      expect(recording.events, ['cli.generate', 'cli.generate']);
+      expect(recording.properties.last['incremental_run_count'], 1);
+      expect(recording.properties.last['incremental_avg_duration_ms'], 400);
+      expect(recording.properties.last['generation_succeeded'], isFalse);
+      expect(
+        recording.properties.last['num_generate_calls'],
+        2,
+        reason: 'The metadata counter increments once per emitted event.',
+      );
+    },
+  );
 }
